Refatoring + Stating that ascription can work as a rename
Previously, this PR claimed that column/field/param names were not checked during ascription. This could be viewed as implying that the column/field/param names given in the type being ascribed have no effect.

This is incorrect. Rather, the child names given in the ascribed type are applied as renames.