Fix YoY growth calculation...
It should divide by previous year sales amount. 

It is correct in this documentation page
```
https://learn.microsoft.com/en-us/dax/best-practices/dax-variables
```

I also assumed the IF check was to prevent dividing by zero, so updated accordingly to check the previous year sales since it is now the denominator. Some quick tests revealed that <blank> would be returned even if divide by zero happened, so the IF check may not be needed.

Finally, I tried to keep the same formatting in the places I made changes, even though they are inconsistent. So more work could be done to clean things up.