Table.FromList includes example with identity splitter
The example is trivial for someone with experience writing M functions.

However, for someone starting writing M, this example provides a "learning by doing" way of defining and using a simple function and solving a simple problem.

If there is a better way of converting a list to a table without splitting by any character, then the documentation could refer to it instead of providing this example. As a beginner myself, I would find it helpful.